Change how system is initialized.
The problem was BARRIERW, that is only called before TM_INIT. As such, we need
to setup the initialization before TM_INIT, through TM_INITs and init_system.

Additionally, we need to setup ID and everything, including barrier setup in
init_system. To do so, introduce sys_init_system, that is called
_AT_THE_BEGINING_ of init_system. This way, we fetch the identifiers from the
platform, and we may call the init_barrier (now, a requirement for platforms).